Update validFor type to uint64

+++ test/utils/sign.go
@@ -60,7 +60,7 @@ func newTimedSignature(taStoreContract *framework.Contract, privateKey *ecdsa.Pr
 	if err != nil {
 		return nil, err
 	}
-	messageHash, signature, err := generateTimedSignature(int64(validFor), privateKey, nonce, targetFunctionHash)
+	messageHash, signature, err := generateTimedSignature(validFor, privateKey, nonce, targetFunctionHash)
 	if err != nil {
 		return nil, fmt.Errorf("failed to generate timed signature: %v", err)
 	}
@@ -86,13 +86,13 @@ func getNonce(taStoreContract *framework.Contract, address common.Address) (uint
 	return nonce, nil
 }
 
-func generateTimedSignature(validFor int64, privateKey *ecdsa.PrivateKey, nonce uint64, targetFunctionHash [32]byte) (messageHash [32]byte, signature []byte, err error) {
+func generateTimedSignature(validFor uint64, privateKey *ecdsa.PrivateKey, nonce uint64, targetFunctionHash [32]byte) (messageHash [32]byte, signature []byte, err error) {
 	address := crypto.PubkeyToAddress(privateKey.PublicKey)
 
 	// Step 1: Create the message hash
 	// Combine validFor timestamp, signer's address, nonce, and targetFunctionHash, then hash with Keccak256
 	messageHash = crypto.Keccak256Hash(
-		common.LeftPadBytes(big.NewInt(validFor).Bytes(), 8),
+		common.LeftPadBytes(big.NewInt(int64(validFor)).Bytes(), 8),
 		common.LeftPadBytes(address.Bytes(), 20),
 		common.LeftPadBytes(big.NewInt(int64(nonce)).Bytes(), 8),
 		targetFunctionHash[:],
